Added deletion tests Created a new unit test class
