+++ ExpressionFrameworkTests/Sacaza Framework/CollectionNodeSacazaTests.swift
@@ -37,25 +37,6 @@ class CollectionNodeSacazaTests: XCTestCase {
         collectionNodeSacaza = CollectionNodeSacaza(collectionNode: collectionNode, dataSource: self, delegate: self, sacaza: sacaza)
     }
     
-    
-    // MARK:- Sacaza tests
-    
-    func testTheAdjustedIndexPaths() {
-        let downloadedPresenceItems =  [
-            PresenceInfo(indexPath: IndexPath(row:3 , section: 0), expressionContainer: TestableAdvert("advertAtRow3")),
-            PresenceInfo(indexPath: IndexPath(row:7 , section: 0), expressionContainer: TestableAdvert("advertAtRow7")),
-            PresenceInfo(indexPath: IndexPath(row:11 , section: 0), expressionContainer: TestableAdvert("advertAtRow11")),
-            PresenceInfo(indexPath: IndexPath(row:15 , section: 0), expressionContainer: TestableAdvert("advertAtRow15"))
-        ]
-        
-        let sacaza = Sacaza(presenceItems: downloadedPresenceItems, numberOfOriginalItems: feedItems.count)
-        
-        XCTAssertEqual(0, sacaza.calculateAdjustedIndexPath(forItemAtOriginalIndexPath: IndexPath(row: 0, section: 0)).row)
-        XCTAssertEqual(4, sacaza.calculateAdjustedIndexPath(forItemAtOriginalIndexPath: IndexPath(row: 3, section: 0)).row)
-        XCTAssertEqual(8, sacaza.calculateAdjustedIndexPath(forItemAtOriginalIndexPath: IndexPath(row: 6, section: 0)).row)
-        XCTAssertEqual(12, sacaza.calculateAdjustedIndexPath(forItemAtOriginalIndexPath: IndexPath(row: 9, section: 0)).row)
-    }
-    
     // MARK:- Collection Node Tests
     
     func testTheNumberOfItems() {
@@ -142,7 +123,33 @@ class CollectionNodeSacazaTests: XCTestCase {
     }
     
     func testDeletions() {
+        // given
+        let deletions = [0,2,4,6,8]
+        let deletionIndexPaths = deletions.map({IndexPath(row: $0, section: 0)})
         
+        // when
+        feedItems.insertItems([], deleteItemsAtIndexes:deletions)
+        collectionNodeSacaza!.insertItems(insertions: [], deletions:deletionIndexPaths)
+        
+        let expectedFeedItems: [Any] = ["two", TestableAdvert("advertAtRow3"), "four", "six", TestableAdvert("advertAtRow7"), "eight", TestableAdvert("advertAtRow11"), "ten"]
+        
+        // then
+        for (index, finalFeedItem) in expectedFeedItems.enumerated() {
+            switch finalFeedItem {
+            case let item as String:
+                guard let cell = collectionNode!.nodeForItem(at: IndexPath(row: index, section: 0)) as? TestableCellNode else {
+                    XCTFail("Expected TestableCellNode type")
+                    return
+                }
+                XCTAssertEqual(cell.text, item)
+            case _ as TestableAdvert:
+                if let _ = collectionNode!.nodeForItem(at: IndexPath(row: index, section: 0)) as? TestableCellNode {
+                    XCTFail("Expected ASCellNode for Advert")
+                }
+            default:
+                XCTFail("Unexpected scenario")
+            }
+        }
     }
     
     func testInsertionsAndDeletions() {
