+++ Sacaza Framework/Swift Extensions/Array+Insertions.swift
@@ -11,7 +11,7 @@ import Foundation
 // TODO we may not need this extension any more and it probably could be deleted
 public extension Array {
     
-    public mutating func insertItems(_ insertedItems: [(Element, Int)], deleteItemsAtIndexes deletedIndexes: [Int]) {
+    mutating func insertItems(_ insertedItems: [(Element, Int)], deleteItemsAtIndexes deletedIndexes: [Int]) {
         var amountToIncrementWIthinAdditions = 0
         insertedItems.forEach({ tuple in
             let index = amountToIncrementWIthinAdditions + tuple.1
